all css done final one

+++ frontEnd/Screens/Owner/OwnerHomePage.jsx
@@ -13,14 +13,17 @@ import {
   TouchableOpacity,
   Image,
   ActivityIndicator,
+  Alert
 } from "react-native";
 import { useNavigation } from "@react-navigation/native";
-import { Ionicons, MaterialIcons, AntDesign } from "@expo/vector-icons";
+import { Ionicons, MaterialIcons,MaterialCommunityIcons,FontAwesome5} from "@expo/vector-icons";
 import { APP_API_URL } from "../../env";
 import SessionStorage from "react-native-session-storage";
 import Toast from "react-native-toast-message";
 
+
 const OwnerHomePage = () => {
+  
   const navigation = useNavigation();
   const [filteredProperties, setFilteredProperties] = useState([]);
   const [properties, setProperties] = useState([]);
@@ -115,46 +118,35 @@ const OwnerHomePage = () => {
   };
 
   const handleDeleteProperty = (propertyId) => {
-    Toast.show({
-      type: 'info',
-      text1: 'Deleting Property',
-      text2: 'Are you sure you want to delete this property?',
-      position: 'bottom',
-      autoHide: false,
-      onPress: () => {
-        setLoading(true);
-        fetch(`${APP_API_URL}/property/delet/${propertyId}`, {
-          method: "DELETE",
-        })
-          .then(() => {
-            setLoading(false);
-            fetchOwnerProperties();
-            Toast.hide();
-            Toast.show({
-              type: 'success',
-              text1: 'Success',
-              text2: 'Property deleted successfully',
-              position: 'bottom',
-              bottomOffset:800,
-            });
-          })
-          .catch((error) => {
-            console.error("Error deleting property:", error);
-            setLoading(false);
-            Toast.hide();
-            Toast.show({
-              type: 'error',
-              text1: 'Error',
-              text2: 'Error deleting property',
-              position: 'bottom',
-              bottomOffset:800,
-            });
-          });
-      },
-      onHide: () => {
-        Toast.hide();
-      },
-    });
+    Alert.alert(
+      "Confirm Delete",
+      "Are you sure you want to delete this property?",
+      [
+        {
+          text: "Cancel",
+          style: "cancel",
+        },
+        {
+          text: "Delete",
+          onPress: () => {
+            setLoading(true);
+            fetch(`${APP_API_URL}/property/delet/${propertyId}`, {
+              method: "DELETE",
+            })
+              .then(() => {
+                setLoading(false);
+                alert("Property deleted successfully!");
+                fetchOwnerProperties();
+              })
+              .catch((error) => {
+                console.error("Error deleting property:", error);
+                setLoading(false);
+              });
+          },
+        },
+      ],
+      { cancelable: false }
+    );
   };
 
   return (
@@ -196,74 +188,76 @@ const OwnerHomePage = () => {
         </View>
         <ScrollView>
           <View>
-            {showFilteredProperties ? (
-              filteredProperties.map((property) => (
-                <View key={property.id} style={styles.propertyItem}>
-                  <TouchableOpacity
-                    onPress={() =>
-                      navigation.navigate("ProductsDetails", {
-                        propertyid: property.id,
-                        userid: userid,
-                      })
-                    }
-                  >
-                    <Image
-                      style={styles.propertyImage}
-                      source={{ uri: property.image[0] }}
-                    />
-                    <View style={styles.propertyDetails}>
-                      <Text style={styles.propertyTitle}>{property.Name}</Text>
-                      <Text style={styles.propertyLocation}>
-                        <MaterialIcons name="location-pin" size={18} color="grey" />
-                        {property.location}
-                      </Text>
-                      <Text style={styles.propertyPrice}>
-                        dt {property.Price} / Visit
-                      </Text>
-                    </View>
-                  </TouchableOpacity>
+          {showFilteredProperties ? (
+          filteredProperties.map((property) => (
+            <View key={property.id} style={styles.propertyItem}>
+              <TouchableOpacity
+                onPress={() =>
+                  navigation.navigate("ProductsDetails", {
+                    propertyid: property.id,
+                    userid: userid,
+                  })
+                }
+              >
+                <Image
+                  style={styles.propertyImage}
+                  source={{ uri: property.image[0] }}
+                />
+                <View style={styles.propertyDetails}>
+                  <Text style={styles.propertyTitle}>{property.Name}</Text>
+                  <Text style={styles.propertyLocation}>
+                    <MaterialIcons name="location-pin" size={18} color="grey" />
+                    <MaterialIcons name="location-pin" size={18} color="grey" />
+                    {property.location}
+                  </Text>
+                  <Text style={styles.propertyPrice}>
+                    dt {property.Price} / Visit
+                  </Text>
                 </View>
-              ))
-            ) : (
-              properties.map((property) => (
-                <View key={property.id} style={styles.propertyItem}>
-                  <TouchableOpacity
-                    onPress={() =>
-                      navigation.navigate("ProductsDetails", {
-                        propertyid: property.id,
-                        // userId: userId,
-                      })
-                    }
-                  >
-                    <Image style={styles.propertyImage} source={{ uri: property.image[0] }} />
-                    <View style={styles.propertyDetails}>
-                      <Text style={styles.propertyTitle}>{property.Name}</Text>
-                      <Text style={styles.propertyLocation}>
-                        <MaterialIcons name="location-pin" size={18} color="grey" />
-                        {property.location}
-                      </Text>
-                      <Text style={styles.propertyPrice}>dt {property.Price} / Visit</Text>
-                    </View>
-                  </TouchableOpacity>
-                  <View style={styles.actionButtons}>
-                    <TouchableOpacity
-                      style={styles.editButton}
-                      onPress={() => navigation.navigate("EditProperty",{propertyId:property.id})}
-                    >
-                      <Text style={styles.buttonText}> <AntDesign name="edit" size={20} color="black" /></Text>
-                    </TouchableOpacity>
-                    <TouchableOpacity
-                      style={styles.deleteButton}
-                      onPress={() => handleDeleteProperty(property.id)}
-                    >
-                      <Text style={styles.buttonTextt}><AntDesign name="delete" size={20} color="black" /></Text>
-                    </TouchableOpacity>
-                  </View>
+              </TouchableOpacity>
+            </View>
+          ))
+        ) : (
+          properties.map((property) => (
+            <View key={property.id} style={styles.propertyItem}>
+              <TouchableOpacity
+                onPress={() =>
+                  navigation.navigate("ProductsDetails", {
+                    propertyid: property.id,
+                    // userId: userId,
+                  })
+                }
+              >
+                <Image style={styles.propertyImage} source={{ uri: property.image[0] }} />
+                <View style={styles.propertyDetails}>
+                  <Text style={styles.propertyTitle}>{property.Name}</Text>
+                  <Text style={styles.propertyLocation}>
+                    <MaterialIcons name="location-pin" size={18} color="grey" />
+                    {property.location}
+                  </Text>
+                  <Text style={styles.propertyPrice}> {property.Price} dt / Visit</Text>
                 </View>
-              ))
-            )}
-          </View>
-        </ScrollView>
+              </TouchableOpacity>
+              <View style={styles.actionButtons}>
+                <TouchableOpacity
+                  style={styles.updateprop}
+                  onPress={() => navigation.navigate("EditProperty",{propertyId:property.id})}
+                >
+                  <Text style={styles.buttonText}><MaterialCommunityIcons name="home-edit-outline" size={24} color="#999999" /></Text>
+                </TouchableOpacity>
+                <TouchableOpacity
+                  style={styles.updateprop}
+                  onPress={() => handleDeleteProperty(property.id)}
+                >
+                  <Text style={styles.buttonTextt}><FontAwesome5 name="trash-alt" size={19} color="#999999" /></Text>
+                </TouchableOpacity>
+              </View>
+            </View>
+          ))
+        )}
+      </View>
+    </ScrollView>
+   
       </View>
     </ScrollView>
   );
@@ -424,7 +418,13 @@ const styles = StyleSheet.create({
     left: 145,
     bottom: '18%',
     marginBottom: 5,
+    
   },
+  updateprop: {
+    marginLeft: 8,
+    
+  },
+ 
 });
 
 export default OwnerHomePage;
